fix syntax on cleanup

+++ hooks/post_gen_project.py
@@ -1,7 +1,7 @@
 import shutil
+create_docs = '{{ cookiecutter.create_docs }}'
 
 if __name__ == "__main__":
-    { % if cookiecutter.create_docs == "yes" - %}
-    shutil.rmtree("docs", ignore_errors=True)
-    { % - endif %}
+    if create_docs != "yes":
+        shutil.rmtree("docs", ignore_errors=True)
     shutil.rmtree("licenses", ignore_errors=True)
